+++ CHANGES.md
@@ -1,3 +1,9 @@
+
+
+- Topkg.String.parse_version. Support for the new OCaml 
+  version string format (https://github.com/ocaml/ocaml/pull/9712)
+
+
 v1.0.4 2021-10-04 Zagreb
 -------------------------